+++ includes/gateways/class-wc-heidelpay-gateway-va.php
@@ -21,15 +21,31 @@ if (!defined('ABSPATH')) {
 
 require_once WC_HEIDELPAY_PLUGIN_PATH . DIRECTORY_SEPARATOR . 'includes' . DIRECTORY_SEPARATOR . 'abstracts' .
     DIRECTORY_SEPARATOR . 'abstract-wc-heidelpay-payment-gateway.php';
+require_once WC_HEIDELPAY_PLUGIN_PATH . DIRECTORY_SEPARATOR . 'includes' . DIRECTORY_SEPARATOR . 'traits' .
+    DIRECTORY_SEPARATOR . 'trait-wc-heidelpay-subscription-gateway.php';
 
 use Heidelpay\PhpPaymentApi\PaymentMethods\PayPalPaymentMethod;
 
 class WC_Gateway_HP_VA extends WC_Heidelpay_Payment_Gateway
 {
+    use WC_Heidelpay_Subscription_Gateway;
+
     /** @var array Array of locales */
     public $locale;
     protected $bookingModes;
 
+    /**
+     * WC_Gateway_HP_VA constructor.
+     */
+    public function __construct()
+    {
+        parent::__construct();
+        $this->constructorAddon();
+    }
+
+    /**
+     * sets the Payment Method.
+     */
     public function setPayMethod()
     {
         $this->payMethod = new PayPalPaymentMethod();
@@ -48,6 +64,7 @@ class WC_Gateway_HP_VA extends WC_Heidelpay_Payment_Gateway
     public function init_form_fields()
     {
         parent::init_form_fields();
+        $this->initFormFieldsAddon();
 
         $this->form_fields['security_sender']['default'] = '31HA07BC8142C5A171745D00AD63D182';
         $this->form_fields['user_login']['default'] = '31ha07bc8142c5a171744e5aef11ffd3';